+++ msal/src/main/java/com/microsoft/identity/nativeauth/statemachine/Error.kt
@@ -21,7 +21,7 @@
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
 
-package com.microsoft.identity.client.statemachine
+package com.microsoft.identity.nativeauth.statemachine
 
 /**
  * Error is a base class for all errors present in the Native Auth.
