+++ msal/src/test/java/com/microsoft/identity/nativeauth/NativeAuthOAuth2ConfigurationTest.kt
@@ -20,9 +20,9 @@
 //  LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 //  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 //  THE SOFTWARE.
-package com.microsoft.identity.client.nativeauth
+package com.microsoft.identity.nativeauth
 
-import com.microsoft.identity.common.java.providers.nativeauth.NativeAuthOAuth2Configuration
+import com.microsoft.identity.common.nativeauth.providers.NativeAuthOAuth2Configuration
 import junit.framework.Assert.assertEquals
 import org.junit.Test
 import org.junit.runner.RunWith
